+++ public/admin/_app/immutable/nodes/18._XaR6E6S.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/B_ox1QAz2.js";export{e as component};