+++ public/admin/_app/immutable/nodes/20.sbTbXPw7.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/UzIuh07e.js";export{e as component};