+++ public/admin/_app/immutable/nodes/22.Cm_oqMv5.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/Ya-rmQyq.js";export{e as component};