+++ public/admin/_app/immutable/nodes/24.DBQbH319.js
@@ -1,0 +1,1 @@
+import{t as e}from"../chunks/cq54g010.js";export{e as component};